+++ CHANGELOG.md
@@ -25,7 +25,8 @@ project adheres to
   `Number::is_integer()`.
 * Changed the error type of `Scope::get_function()`.
 * Added a variant to `ScopeError`.
-* `sass::Value::Variable` now holds a `Name` rather than just a `String`.
+* The sass `Value::Variable` and `Item::VariableDeclaration` variants
+  now holds a `Name` rather than just a `String` for the variable name.
 
 ### Improvements